Agent3-9/summarizer agent/requirements.txt
Crewai>=0.51.0
litellm>=1.43.0
tavily-python>=0.3.5
google-generativeai>=0.8.3
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
pydantic>=2.8.0
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn>=0.30.0
